requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic-settings
pytest
psycopg2-binary